refactor: move handleResponse logic into MessageQueue
